+++ src/components/ContextMenu/ContextMenu.tsx
@@ -1,22 +1,14 @@
 import React from 'react';
-import { Menu, Popup, StrictPopupProps } from 'semantic-ui-react';
+import { Menu, Popup } from 'semantic-ui-react';
 import { IAction, IActionParamForRun } from './IAction';
 import { TestsAreDemoCheat, createTestids } from '@famiprog-foundation/tests-are-demo';
 import _ from 'lodash';
 
 export type Point = { x: number, y: number };
-
-type Position = StrictPopupProps["position"];
-
-type IParamsForAction = {
-  selection: any[];
-  position?: Position;
-  [key: string]: any;
-}
 interface ContextMenuProps {
   actions: IAction[];
 
-  paramsForAction: IParamsForAction;
+  paramsForAction: IActionParamForRun;
   /**
    * if undefined => the menu is closed else {x, y} position where the menu should open
    */